Move update_training_job tests out of test_recreate.py
test_recreate.py is about job recreation, and it lives in truss-train/tests/
even though it exercises truss.cli.train.core. TestUpdateTrainingJob moves to
truss/tests/cli/train/test_job_update.py, which names what it tests and sits in
the tree for the package under test.

All five update assertions stay together in the new file, so a change to the
call signature still can't pass one suite while breaking another.
